requirements.txt
annotated-types==0.7.0
anthropic==0.75.0
anyio==4.11.0
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
distro==1.9.0
docstring_parser==0.17.0
evdev==1.9.2
exceptiongroup==1.3.1
google-api-core==2.28.1
google-auth==2.43.0
google-cloud-speech==2.34.0
google-cloud-texttospeech==2.33.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.12.0
numpy==2.2.6
pillow==12.0.0
proto-plus==1.26.1
protobuf==6.33.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pygame==2.6.1
pynput==1.8.1
pystray==0.19.5
python-xlib==0.33
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
scipy==1.15.3
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
